requirements.txt
# MCP
mcp>=1.6.0

# LLM via OpenRouter
openai>=1.30.0

# Google APIs (Gmail + Calendar)
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.131.0

# Redis (async)
redis[asyncio]>=5.0.3

# Vector DB + embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# FastAPI + ASGI server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Settings management
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# HTTP client (MCP server health probe)
httpx>=0.27.0

# Twilio (phone calls)
twilio>=9.1.0

# Structured logging
structlog>=24.2.0
colorama>=0.4.6

# Utilities
python-multipart>=0.0.9
